# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для сокращения ссылок",
        "title": "Url-shortener",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/url": {
            "post": {
                "description": "Сохраняет оригинальный URL и возвращает alias. Если alias не передан, генерируется случайный.",
                "tags": [
                    "URL"
                ],
                "summary": "Сохраняет URL",
                "parameters": [
                    {
                        "description": "Данные запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Request"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "URL успешно сохранён",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный JSON",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/url/{alias}": {
            "delete": {
                "description": "Удаляет URL по alias. Возвращает статус 204 No Content в случае успеха.",
                "tags": [
                    "URL"
                ],
                "summary": "Удаляет URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alias для удаления",
                        "name": "alias",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "URL успешно удалён"
                    },
                    "422": {
                        "description": "Alias не указан",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка удаления URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/{alias}": {
            "get": {
                "description": "Получает alias из URL, ищет соответствующий оригинальный URL и выполняет редирект (HTTP 302).",
                "tags": [
                    "URL"
                ],
                "summary": "Перенаправляет на оригинальный URL по alias",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alias для редиректа",
                        "name": "alias",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirected to original URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Request": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "alias": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}