# internal/http-server/handlers/model/model.go
package model

import resp "url-shortener/internal/lib/api/response"

type Request struct {
	URL   string `json:"url" validate:"required,url"`
	Alias string `json:"alias,omitempty"`
}

type Response struct {
	resp.Response
	Alias string `json:"alias,omitempty"`
}